feat(moves candidate components to a new package in the repo root): moves candidate components
