+++ packages/styled-components/components/common/html/themes/default.js
