+++ packages/styled-components/components/common/link/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import parseUrl from 'utils/getRelativeUrl';
-import history from 'utils/history';
-import withThemes from 'component-candidates/hoc/withThemes';
+import parseUrl from '@irvingjs/core/utils/getRelativeUrl';
+import history from '@irvingjs/core/utils/history';
+import withThemes from 'components/hoc/withThemes';
 import * as defaultStyles from './themes/default';
 
 /**
