+++ packages/styled-components/components/layouts/app/component.json
