+++ packages/styled-components/components/layouts/fragment/themes/default.js
