+++ packages/styled-components/components/modules/logo/component.json
