+++ packages/styled-components/components/modules/logo/index.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import React from 'react';
 import PropTypes from 'prop-types';
-import withThemes from 'component-candidates/hoc/withThemes';
+import withThemes from 'components/hoc/withThemes';
 import * as defaultStyles from './themes/default';
 
 /**
