+++ packages/styled-components/components/modules/logo/index.php
