+++ packages/styled-components/components/modules/logo/themes/default.js
