+++ packages/styled-components/components/modules/menu/themes/defaultVertical.js
