+++ packages/styled-components/components/wordpress/page/archive-title/component.json
