+++ packages/styled-components/components/wordpress/post/byline/themes/default.js
