+++ packages/styled-components/components/wordpress/post/excerpt/component.json
