+++ packages/styled-components/components/wordpress/post/featuredMedia/component.json
