+++ packages/styled-components/components/wordpress/post/featuredMedia/index.php
