+++ packages/styled-components/components/wordpress/post/list/component.json
