+++ packages/styled-components/components/wordpress/post/permalink/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Link from 'component-candidates/common/link';
+import Link from 'components/common/link';
 
 /**
  * Post permalink.
