+++ packages/styled-components/package.json
@@ -1,0 +1,49 @@
+{
+	"name": "@irvingjs/styled-components",
+	"version": "2.11.0-beta.17",
+	"description": "Integration for React components with attached styles for Irving",
+	"files": [
+		"components/**/*",
+		"config/**/*",
+		"irving.config.server.js",
+		"irving.config.js",
+		"!*.test.js"
+	],
+	"main": "index.js",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/alleyinteractive/irving/packages/styled-components"
+	},
+	"author": "Alley (Alley.co)",
+	"license": "MIT",
+	"scripts": {
+		"test": "jest",
+		"test:watch": "jest --watch"
+	},
+	"dependencies": {
+		"babel-plugin-styled-components": "^1.10.7",
+		"he": "^1.2.0",
+		"react": "^16.13.1",
+		"react-dom": "^16.13.1",
+		"styled-components": "^5.1.0",
+		"stylelint": "^13.3.3",
+		"stylelint-config-styled-components": "^0.1.1",
+		"stylelint-custom-processor-loader": "^0.6.0",
+		"stylelint-order": "^4.0.0",
+		"stylelint-processor-styled-components": "^1.10.0"
+	},
+	"devDependencies": {
+		"@irvingjs/babel-preset-irving": "^2.11.0-beta.17",
+		"@irvingjs/core": "^2.11.0-beta.17",
+		"babel-plugin-module-resolver": "^4.0.0",
+		"enzyme": "^3.11.0",
+		"jest": "^25.5.3"
+	},
+	"peerDependencies": {
+		"@irvingjs/core": ">=2.0.0"
+	},
+	"publishConfig": {
+		"access": "public"
+	},
+	"gitHead": "2b57a99dc78a57506cc60cf709b756d1cc5ecef5"
+}
